+++ batch/src/main/kotlin/com/nexters/newsletterfeeder/NewsletterFeederApplication.kt
@@ -3,6 +3,7 @@ package com.nexters.newsletterfeeder
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.context.properties.ConfigurationPropertiesScan
 import org.springframework.boot.runApplication
+import org.springframework.context.annotation.ComponentScan
 import org.springframework.integration.annotation.IntegrationComponentScan
 import org.springframework.scheduling.annotation.EnableScheduling
 
@@ -10,6 +11,7 @@ import org.springframework.scheduling.annotation.EnableScheduling
 @EnableScheduling
 @IntegrationComponentScan
 @ConfigurationPropertiesScan
+@ComponentScan(basePackages = ["com.nexters.newsletterfeeder", "com.nexters.external"])
 class NewsletterFeederApplication
 
 fun main(args: Array<String>) {
