+++ gradle/libs.version.toml
@@ -13,6 +13,7 @@ jackson = "2.17.0"
 jakarta-activation = "2.1.2"
 junit = "5.10.0"
 google-genai = "1.0.0"
+gson = "2.10.1"
 mockk = "1.13.8"
 
 # Code Quality
@@ -46,6 +47,9 @@ jakarta-activation-api = { group = "jakarta.activation", name = "jakarta.activat
 
 # Google  AI
 google-genai = { group = "com.google.genai", name = "google-genai", version.ref = "google-genai" }
+
+# Gson
+gson = { group = "com.google.code.gson", name = "gson", version.ref = "gson" }
 
 # Test dependencies
 spring-boot-starter-test = { group = "org.springframework.boot", name = "spring-boot-starter-test" }
